Fix #2507: Getting 401 unauthorised error for retrieve permissions API
