+++ catalog-rest-service/src/main/java/org/openmetadata/catalog/CatalogApplication.java
@@ -52,6 +52,7 @@ import org.openmetadata.catalog.exception.ConstraintViolationExceptionMapper;
 import org.openmetadata.catalog.exception.JsonMappingExceptionMapper;
 import org.openmetadata.catalog.resources.CollectionRegistry;
 import org.openmetadata.catalog.resources.config.ConfigResource;
+import org.openmetadata.catalog.resources.permissions.PermissionsResource;
 import org.openmetadata.catalog.resources.search.SearchResource;
 import org.openmetadata.catalog.security.AuthenticationConfiguration;
 import org.openmetadata.catalog.security.Authorizer;
@@ -170,8 +171,10 @@ public class CatalogApplication extends Application<CatalogApplicationConfig> {
       ContainerRequestFilter filter = NoopFilter.class.getConstructor().newInstance();
       environment.jersey().register(filter);
     }
-    // Registering config api
+    // Register config API
     environment.jersey().register(new ConfigResource(catalogConfig, authorizer));
+    // Register permissions API
+    environment.jersey().register(new PermissionsResource(authorizer));
   }
 
   private void registerEventFilter(CatalogApplicationConfig catalogConfig, Environment environment, Jdbi jdbi) {
