+++ catalog-rest-service/src/main/java/org/openmetadata/catalog/resources/config/ConfigResource.java
@@ -21,17 +21,13 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.SecurityContext;
 import lombok.NonNull;
 import org.openmetadata.catalog.CatalogApplicationConfig;
 import org.openmetadata.catalog.resources.Collection;
 import org.openmetadata.catalog.security.AuthenticationConfiguration;
 import org.openmetadata.catalog.security.Authorizer;
 import org.openmetadata.catalog.security.AuthorizerConfiguration;
-import org.openmetadata.catalog.security.Permissions;
-import org.openmetadata.catalog.security.SecurityUtil;
 
 @Path("/v1/config")
 @Api(value = "Get configuration")
@@ -88,20 +84,5 @@ public class ConfigResource {
       authorizerConfiguration = catalogApplicationConfig.getAuthorizerConfiguration();
     }
     return authorizerConfiguration;
-  }
-
-  @GET
-  @Path(("/permissions"))
-  @Operation(
-      summary = "Retrieves permissions for logged in user",
-      tags = "general",
-      responses = {
-        @ApiResponse(
-            responseCode = "200",
-            description = "Permissions for logged in user",
-            content = @Content(mediaType = "application/json", schema = @Schema(implementation = Permissions.class)))
-      })
-  public Permissions getPermissions(@Context SecurityContext securityContext) {
-    return new Permissions(authorizer.listPermissions(SecurityUtil.getAuthenticationContext(securityContext), null));
   }
 }
